Fix & Refactor: Syncing active nav tab with active section by refactoring with addition of Section class
A new class 'Section' is introduced to encapsulate the logic of handling a section, its corresponding nav item and their active status.

The global array SECTIONS that represented all the sections in the page has been refactored to become a Map javascript object where the key of each entry is the section ID, and the value is a Section object that holds the HTML section and anchor elements.

With such refactoring, the section and nav status are now synced properly. Also, a lot of the operations that included searching for a section in the array or traversing the DOM using querySelector to get the anchor or the section, all these operations have been replaced with constant time operations O(1) exploiting the performance benefits of the Map JS object.